+++ src/test/java/com/chriscarini/jetbrains/github/issue/navigation/TestGitHubIssueNavigationVcsRepositoryMappingListenerTest.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 
-public class TestGitHubIssueNavigationStartupActivityTest extends LightPlatformTestCase {
+public class TestGitHubIssueNavigationVcsRepositoryMappingListenerTest extends LightPlatformTestCase {
     @Override
     public void setUp() throws Exception {
         super.setUp();
@@ -26,25 +26,25 @@ public class TestGitHubIssueNavigationStartupActivityTest extends LightPlatformT
         IssueNavigationConfiguration.getInstance(getProject()).setLinks(List.of(myIssueNavigationLink));
 
         // expect
-        assert GitHubIssueNavigationStartupActivity.existsInIssueNavConfig(getProject(), expectedUrl);
-        assert !GitHubIssueNavigationStartupActivity.existsInIssueNavConfig(getProject(), unexpectedUrl);
+        assert GitHubIssueNavigationVcsRepositoryMappingListener.existsInIssueNavConfig(getProject(), expectedUrl);
+        assert !GitHubIssueNavigationVcsRepositoryMappingListener.existsInIssueNavConfig(getProject(), unexpectedUrl);
     }
 
     @Test
     public void testAddNewIssueNavigationConfiguration() {
         // given
         final String expectedUrl = "MY_FAKE_URL";
-        assert IssueNavigationConfiguration.getInstance(getProject()).getLinks().size() == 0;
+        assert IssueNavigationConfiguration.getInstance(getProject()).getLinks().isEmpty();
 
         // when
-        GitHubIssueNavigationStartupActivity.addNewIssueNavigationConfiguration(getProject(), expectedUrl);
+        GitHubIssueNavigationVcsRepositoryMappingListener.addNewIssueNavigationConfiguration(getProject(), expectedUrl);
 
         // then
         final List<IssueNavigationLink> issues = IssueNavigationConfiguration.getInstance(getProject())
             .getLinks()
             .stream()
             .filter(issueNavigationLink -> issueNavigationLink.getLinkRegexp().contains(expectedUrl))
-            .collect(Collectors.toList());
+            .toList();
         assertEquals(1, issues.size());
     }
 }