+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 - Upgrading IntelliJ from 2025.1 to 2025.1.1
+- Migrate from `StartupActivity` (obsolete) -> `VcsRepositoryMappingListener`
 
 ### Deprecated