+++ CHANGELOG.md
@@ -5,10 +5,8 @@
 ## [Unreleased]
 
 ### Added
-- Add more patterns for matching issues (`#123`, `GH-123`, and `ChrisCarini/automatic-github-issue-navigation-configuration-jetbrains-plugin#123`)
 
 ### Changed
-- Upgrading IntelliJ from 2025.1.2 to 2025.1.3
 
 ### Deprecated
 
@@ -17,6 +15,16 @@
 ### Fixed
 
 ### Security
+
+## [3.0.2] - 2025-07-06
+
+### Added
+
+- Add more patterns for matching issues (`#123`, `GH-123`, and `ChrisCarini/automatic-github-issue-navigation-configuration-jetbrains-plugin#123`)
+
+### Changed
+
+- Upgrading IntelliJ from 2025.1.2 to 2025.1.3
 
 ## [3.0.1] - 2025-07-06
 
@@ -315,7 +323,8 @@
 
 - Initial release.
 
-[Unreleased]: https://github.com/ChrisCarini/automatic-github-issue-navigation-configuration-jetbrains-plugin/compare/v3.0.1...HEAD
+[Unreleased]: https://github.com/ChrisCarini/automatic-github-issue-navigation-configuration-jetbrains-plugin/compare/v3.0.2...HEAD
+[3.0.2]: https://github.com/ChrisCarini/automatic-github-issue-navigation-configuration-jetbrains-plugin/compare/v3.0.1...v3.0.2
 [3.0.1]: https://github.com/ChrisCarini/automatic-github-issue-navigation-configuration-jetbrains-plugin/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/ChrisCarini/automatic-github-issue-navigation-configuration-jetbrains-plugin/compare/v2.2.8...v3.0.0
 [2.2.8]: https://github.com/ChrisCarini/automatic-github-issue-navigation-configuration-jetbrains-plugin/compare/v2.2.7...v2.2.8